testing

+++ test_uv_a/main.cpp
@@ -1,21 +1,29 @@
 ﻿#include "xx_uv.h"
 
 int main(int argc, char* argv[]) {
+	if (argc < 2) {
+		std::cout << "need args: port\n";
+		return -1;
+	}
 	xx::UvLoop loop;
-	auto listener = xx::TryMake<xx::UvUdpKcpListener<>>(loop, "0.0.0.0", 12345);
-	assert(listener);
+	auto listener = xx::Make<xx::UvUdpKcpListener<>>(loop, "0.0.0.0", std::atoi(argv[1]));
 	listener->OnAccept = [&loop](auto& peer) {
 		xx::Cout(peer->kcp->conv, " accepted.\n");
 		auto timeouter = xx::TryMake<xx::UvTimer>(loop, 2000, 0);	// 两秒内没数据的连接杀掉
-		timeouter->OnFire = [timeouter, peer] {	// mem holder
+		timeouter->OnFire = [timeouter, peer] {
 			if (!peer->Disposed()) {
 				peer->Dispose();
 				xx::Cout("timeout. peer dispose\n");
 			}
 			timeouter->OnFire = nullptr;	// kill timer
 		};
-		peer->OnReceivePush = [timeouter, peer](xx::Object_s&& msg)->int {		// hold memory
+		peer->OnReceivePush = [timeouter, peer](xx::Object_s&& msg)->int {
 			if (int r = peer->SendPush(msg)) return r;	// echo
+			//peer->Flush();
+			return timeouter->Restart();
+		};
+		peer->OnReceiveRequest = [timeouter, peer](int const& serial, xx::Object_s&& msg)->int {
+			if (int r = peer->SendResponse(serial, msg)) return r;	// echo
 			//peer->Flush();
 			return timeouter->Restart();
 		};
